+++ lib/serialization/cooserializer.js
