+++ lib/serialization/jsserializer.js
