+++ lib/serialization/serializer.js
