+++ lib/storage/adapters/couchbase/cbstoreadapter.js
