+++ lib/storage/adapters/mock/mockstoreadapter.js
