+++ lib/storage/adapters/storeadapter.js
