+++ lib/storage/codecs/codec.js
@@ -1,0 +1,32 @@
+'use strict';
+
+/**
+ * @interface Codec
+ */
+function Codec() {
+}
+
+/**
+ *
+ * @function
+ * @name Codec#serialize
+ * @param {ModelInstance} mdlInst
+ * @returns {*}
+ */
+
+/**
+ *
+ * @function
+ * @name Codec#decode
+ * @param {Schema} schema
+ * @param {*} data
+ * @returns {*}
+ */
+
+/**
+ *
+ * @function
+ * @name Codec#filter
+ * @param {ottoexpr.Expression} expr
+ * @param {Model} model
+ */
